+++ umami/tests/unit/input_vars_tools/test_input_vars_tools.py
@@ -9,7 +9,6 @@ from matplotlib.testing.compare import compare_images
 
 from umami.configuration import logger, set_log_level
 from umami.input_vars_tools.PlottingFunctions import (
-    check_kwargs_var_plots,
     plot_input_vars_jets,
     plot_input_vars_trks,
     plot_n_tracks_per_jet,
@@ -17,46 +16,6 @@ from umami.input_vars_tools.PlottingFunctions import (
 from umami.tools import yaml_loader
 
 set_log_level(logger, "DEBUG")
-
-
-class KwargsCheck_TestCase(unittest.TestCase):
-    def setUp(self):
-        """
-        Duplicating default dict from function.
-        """
-        self.default_kwargs = {
-            "use_atlas_tag": True,
-            "apply_atlas_style": True,
-            "atlas_first_tag": "Internal Simulation",
-            "atlas_second_tag": "$\\sqrt{s}$ = 13 TeV, $t\\bar{t}$ PFlow Jets",
-            "y_scale": 1,
-            "figsize": (6.8, 5),
-            "logy": True,
-            "ylabel": "Normalised Number of Tracks",
-            "leg_fontsize": 10,
-            "leg_ncol": 1,
-            "leg_loc": "upper right",
-            "bin_width_in_ylabel": True,
-            "plot_type": "pdf",
-            "transparent": True,
-            "norm": True,
-        }
-
-    def test_empty_input(self):
-        kwargs = check_kwargs_var_plots({})
-        self.assertEqual(kwargs, self.default_kwargs)
-
-    def test_one_change(self):
-        kwargs = check_kwargs_var_plots({"use_atlas_tag": False})
-        # TODO: change syntax in python 3.9
-        self.default_kwargs.update({"use_atlas_tag": False})
-        self.assertEqual(kwargs, self.default_kwargs)
-
-    def test_custom_default(self):
-        kwargs = check_kwargs_var_plots({"apply_atlas_style": True}, leg_fontsize=99)
-        # TODO: change syntax in python 3.9
-        self.default_kwargs.update({"apply_atlas_style": True, "leg_fontsize": 99})
-        self.assertEqual(kwargs, self.default_kwargs)
 
 
 class JetPlotting_TestCase(unittest.TestCase):
@@ -219,8 +178,6 @@ class JetPlotting_TestCase(unittest.TestCase):
         # Change type in plotting_config to string to produce error
         plotting_config["binning"]["dr"] = "test"
 
-        del plotting_config["plot_settings"]["ratio_cut"]
-
         with self.assertRaises(ValueError):
             plot_input_vars_trks(
                 datasets_filepaths=filepath_list,
@@ -240,8 +197,6 @@ class JetPlotting_TestCase(unittest.TestCase):
         filepath_list = [self.r21_test_file]
         tracks_name_list = ["tracks"]
         labels_list = ["R21 Test"]
-
-        del plotting_config["plot_settings"]["ratio_cut"]
 
         plot_input_vars_trks(
             datasets_filepaths=filepath_list,